refactor(domain): move generic formatters out of submissionRate to format
