+++ packages/app/src/modules/domain/shared/submissionRate.ts
@@ -1,25 +1,13 @@
-export const NARROW_NBSP = " ";
-
-export function roundOneDecimal(value: number): number {
-	return Math.round(value * 10) / 10;
-}
+import {
+	formatCount,
+	formatRate,
+	NARROW_NBSP,
+	roundOneDecimal,
+} from "./format";
 
 export function computeRate(submitted: number, obligated: number): number {
 	if (obligated === 0) return 0;
 	return roundOneDecimal((submitted / obligated) * 100);
-}
-
-export function formatPointsAbs(points: number): string {
-	const rounded = roundOneDecimal(Math.abs(points));
-	return rounded.toFixed(1).replace(".", ",");
-}
-
-export function formatRate(rate: number): string {
-	return rate.toFixed(1).replace(".", ",");
-}
-
-export function formatCount(count: number): string {
-	return count.toLocaleString("fr-FR").replace(/ /g, NARROW_NBSP);
 }
 
 type CampaignRateData = {
